+++ R/xmile_helpers.R
@@ -108,6 +108,7 @@ sanitise_aux_equation <- function(equation, vendor) {
     stringr::str_replace_all("\n|\t|~| ","") %>%
     stringr::str_replace_all("\\{.*?\\}", "") %>%  # removes commentaries
     translate_extrema() %>%
+    translate_math_funs() %>%
     translate_logical_operators(vendor) %>%
     translate_comparison_operators() %>%
     translate_time_builtins() %>%